+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add faster build scripts for testing (`build:fast`, `build:portable`, `build:nsis`, `build:renderer`)
 - Open in new tab works
 - Open in new window works
+- Add search engine settings
 
 ## v0.8.1 - 2026-02-17